fix(versioning): write specific versions to packages on release
NPM is unable to evaluate correct versions with ^ on tags, as they have no semver logic to them, it
seems. Nevertheless we want specific versions anyways as change to child package indeed makes change
to parent packages in regards of versions. So being strict about this relationship is a good thing